refactor: sse test file
