+++ transport/sse_test.go
@@ -5,8 +5,11 @@ import (
 	"log"
 	"net"
 	"net/http"
+	"net/url"
 	"testing"
 	"time"
+
+	"github.com/ThinkInAIXYZ/go-mcp/pkg"
 )
 
 func TestSSE(t *testing.T) {
@@ -100,4 +103,120 @@ func getAvailablePort() (int, error) {
 
 	port := addr.Addr().(*net.TCPAddr).Port
 	return port, nil
+}
+
+func Test_joinPath(t *testing.T) {
+	type args struct {
+		u    *url.URL
+		elem []string
+	}
+	tests := []struct {
+		name string
+		args args
+		want string
+	}{
+		{
+			name: "1",
+			args: args{
+				u: func() *url.URL {
+					uri, err := url.Parse("https://google.com/api/v1")
+					if err != nil {
+						panic(err)
+					}
+					return uri
+				}(),
+				elem: []string{"/test"},
+			},
+			want: "https://google.com/api/v1/test",
+		},
+		{
+			name: "2",
+			args: args{
+				u: func() *url.URL {
+					uri, err := url.Parse("/api/v1")
+					if err != nil {
+						panic(err)
+					}
+					return uri
+				}(),
+				elem: []string{"/test"},
+			},
+			want: "/api/v1/test",
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			joinPath(tt.args.u, tt.args.elem...)
+			if got := tt.args.u.String(); got != tt.want {
+				t.Errorf("joinPath() = %v, want %v", got, tt.want)
+			}
+		})
+	}
+}
+
+func Test_sseClientTransport_handleSSEEvent(t1 *testing.T) {
+	type fields struct {
+		serverURL *url.URL
+		logger    pkg.Logger
+	}
+	type args struct {
+		event string
+		data  string
+	}
+	tests := []struct {
+		name   string
+		fields fields
+		args   args
+		want   string
+	}{
+		{
+			name: "1",
+			fields: fields{
+				serverURL: func() *url.URL {
+					uri, err := url.Parse("https://api.baidu.com/mcp")
+					if err != nil {
+						panic(err)
+					}
+					return uri
+				}(),
+				logger: pkg.DefaultLogger,
+			},
+			args: args{
+				event: "endpoint",
+				data:  "/sse/messages",
+			},
+			want: "https://api.baidu.com/sse/messages",
+		},
+		{
+			name: "2",
+			fields: fields{
+				serverURL: func() *url.URL {
+					uri, err := url.Parse("https://api.baidu.com/mcp")
+					if err != nil {
+						panic(err)
+					}
+					return uri
+				}(),
+				logger: pkg.DefaultLogger,
+			},
+			args: args{
+				event: "endpoint",
+				data:  "https://api.google.com/sse/messages",
+			},
+			want: "https://api.google.com/sse/messages",
+		},
+	}
+	for _, tt := range tests {
+		t1.Run(tt.name, func(t1 *testing.T) {
+			t := &sseClientTransport{
+				serverURL:    tt.fields.serverURL,
+				logger:       tt.fields.logger,
+				endpointChan: make(chan struct{}),
+			}
+			t.handleSSEEvent(tt.args.event, tt.args.data)
+			if t.messageEndpoint.String() != tt.want {
+				t1.Errorf("handleSSEEvent() = %v, want %v", t.messageEndpoint.String(), tt.want)
+			}
+		})
+	}
 }
